Tic Tac Toe Advanced
This is a complete Tic tac toe game, it includes many functionalities, you can play games
against another human, you can play against the computer, you can also even let the computer
play against itself. The A.I is very good, actualy it is unbeatable, the most you can do is to
draw the games. Also, the game include statistics (number of win, draw an lost games) for the
current match between players. The A.I for this tictactoe game is an implementation of the